fix(data): address review on OpenR1 dataset
- Rename `config_name` to `subset` so a YAML `subset:` is honored
  instead of being silently swallowed by **kwargs (matches the
  ResponseDataset / ResponseDatasetConfig convention).
- Move the OpenR1-specific test out of test_response_dataset_registry.py
  (scoped to registry dispatch) into a dedicated
  tests/unit/data/datasets/test_openr1_math_dataset.py, and add a test
  asserting the default `subset` ('default') is loaded.

Addresses @RayenTian's review comments.
